Adding vue-fuse (again)
Thanks for the great feedback on the initial PR (https://github.com/vuejs/vue-curated/pull/30)

I've updated the library, and should now meet the criteria laid out in the contribution guidelines. I know that this collection is reserved for a few essential plugins, and therefore am not 100% sure of `vue-fuse` qualifies, but I figured I'd let y'all decide.

Cheers!